refactor common_ast
