+++ common/ast/src/ast/query.rs
@@ -15,9 +15,9 @@
 use std::fmt::Display;
 use std::fmt::Formatter;
 
-use crate::parser::ast::display_identifier_vec;
-use crate::parser::ast::Expr;
-use crate::parser::ast::Identifier;
+use crate::ast::display_identifier_vec;
+use crate::ast::Expr;
+use crate::ast::Identifier;
 
 // Root node of a query tree
 #[derive(Debug, Clone, PartialEq)]
