+++ common/ast/tests/it/ast.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common_ast::parser::ast::*;
+use common_ast::ast::*;
 use pretty_assertions::assert_eq;
 
 #[test]
