+++ common/ast/tests/it/main.rs
@@ -13,6 +13,6 @@
 // limitations under the License.
 
 mod ast;
-mod rule;
+mod parser;
 mod token;
 mod udfs;
