+++ query/src/sql/statements/analyzer_expr.rs
@@ -16,9 +16,9 @@ use std::convert::TryFrom;
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use common_ast::parser::expr::ExprTraverser;
-use common_ast::parser::expr::ExprVisitor;
 use common_ast::udfs::UDFDefinition;
+use common_ast::udfs::UDFExprTraverser;
+use common_ast::udfs::UDFExprVisitor;
 use common_ast::udfs::UDFFetcher;
 use common_ast::udfs::UDFParser;
 use common_ast::udfs::UDFTransformer;
@@ -497,7 +497,7 @@ impl ExprRPNBuilder {
             context,
             rpn: Vec::new(),
         };
-        ExprTraverser::accept(expr, &mut builder).await?;
+        UDFExprTraverser::accept(expr, &mut builder).await?;
         Ok(builder.rpn)
     }
 
@@ -665,7 +665,7 @@ impl UDFFetcher for ExprRPNBuilder {
 }
 
 #[async_trait]
-impl ExprVisitor for ExprRPNBuilder {
+impl UDFExprVisitor for ExprRPNBuilder {
     async fn pre_visit(&mut self, expr: &Expr) -> Result<Expr> {
         if let Expr::Function(function) = expr {
             if !is_builtin_function(&function.name.to_string()) {
